Removal of old terminal form. Cleanup. Removal of console.log
