+++ src/client/src/components/TerminalForm/TerminalBaseForm.styled.ts
