+++ src/client/src/components/TerminalForm/TerminalFormRoutes.tsx
@@ -1,10 +1,10 @@
 import { RouteObject } from "react-router-dom";
-import TerminalForm2 from "./TerminalForm2";
+import TerminalForm from "./TerminalForm";
 
 export const terminalFormBasePath = "form/terminal";
 
 export const terminalFormRoutes: RouteObject[] = [
-  { path: terminalFormBasePath, element: <TerminalForm2 /> },
-  { path: `${terminalFormBasePath}/clone/:id`, element: <TerminalForm2 mode={"clone"} /> },
-  { path: `${terminalFormBasePath}/edit/:id`, element: <TerminalForm2 mode={"edit"} /> },
+  { path: terminalFormBasePath, element: <TerminalForm /> },
+  { path: `${terminalFormBasePath}/clone/:id`, element: <TerminalForm mode={"clone"} /> },
+  { path: `${terminalFormBasePath}/edit/:id`, element: <TerminalForm mode={"edit"} /> },
 ];
